requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
transformers>=4.12.0
torch>=1.9.0
pydantic>=1.8.0
PyPDF2>=1.26.0
python-docx>=0.8.11
spacy>=3.2.0
numpy>=1.21.0